moved MainView and updated pathing
